split tasks + fix bug
