+++ src/containers/Main/Main.js
@@ -1,53 +1,42 @@
 import React from 'react';
 import './Main.css';
-import {catalog} from '../../utilities/catalog';
-import {getProperties, getRandomElements, getDiscount} from '../../utilities/utilities';
-import {FaTh, FaThLarge} from 'react-icons/fa'
-
-let currentDiscountedElements = getRandomElements(getProperties(catalog, 'title'));
-let currentDiscount = Math.floor(Math.random() * 20) <= 10 ? 10 : 20
+import { DisplayCatalog, currentDiscountedElements, currentDiscount } from '../../components/DisplayCatalog/DisplayCatalog';
+import { FaTh, FaThLarge } from 'react-icons/fa';
 
 class Main extends React.Component {
-    constructor(props){
+    constructor(props) {
         super(props)
         this.state = {
             display: 'normal'
         }
     }
-    displayProducts(size){
+    displayProducts(size) {
         this.setState(() => ({
-           display: size
+            display: size
         }))
-
     }
 
-    render(){
-        let displayCatalog = this.props.catalog.map(element => {
-          let card = <div key={`${element.author}_${element.year}`} className={this.state.display === 'normal' ? 'card' : 'card large'}> 
-                <img src={element.img} alt={element.title} key={element.img.match(/.{5}(?=.jpg)/)[0]} 
-                onClick={this.props.showDescription.bind(this, element)} title='view description'/> 
-                <h2 key={element.title}>{element.title}</h2>
-                <h3 key={element.author}>{element.author}</h3>              
-               <div key={element.price + element.title}>{getDiscount(currentDiscountedElements, element.title, element.price, currentDiscount, this.props.onClick.bind(this, element.title))}</div>
-            </div>    
-            return card
-        }) 
+    render() {
         return (
             <div class='main-container'>
-                 <div className='description-container'>
-                 <h1>Books for life</h1>
-                 <h2>Explore our range of books</h2>
-                 </div>
-                 <div id='books-container' >
-                 <div className='display-wrapper'>   
-                   <FaTh id='normal-display' onClick={this.displayProducts.bind(this, 'normal')} title='default display'/> 
-                   <FaThLarge id='large-display' onClick={this.displayProducts.bind(this, 'large')} title='large display'/>
-                 </div>
-                 {displayCatalog}
-                 </div>
+                <div className='description-container'>
+                    <h1>Books for life</h1>
+                    <h2>Explore our range of books</h2>
+                </div>
+                <div id='books-container' >
+                    <div className='display-wrapper'>
+                        <FaTh id='normal-display' onClick={this.displayProducts.bind(this, 'normal')} title='default display' />
+                        <FaThLarge id='large-display' onClick={this.displayProducts.bind(this, 'large')} title='large display' />
+                    </div>
+                    <DisplayCatalog
+                        display={this.state.display}
+                        showDescription={this.props.showDescription}
+                        click={this.props.click}
+                    />
+                </div>
             </div>
         )
     }
 }
 
-export {Main, currentDiscount, currentDiscountedElements}
+export { Main, currentDiscount, currentDiscountedElements }
